# Speed up write performance